requirements.txt
<<<<<<< HEAD
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
Flask==1.1.1
functools32==3.2.3.post2
futures==3.3.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.4.0
ipaddress==1.0.17
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
keepnote==0.7.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.6
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.5
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycurl==7.43.0.2
Pygments==2.5.2
PyGObject==3.32.0
pyliblzma==0.5.3
pyrsistent==0.16.0
pysqlite==1.0.1
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
rpm==4.14.2.1
scandir==1.10.0
scipy==1.2.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simple-crypt==4.1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.5
stevedore==1.32.0
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.3
typing==3.7.4.1
unity-lens-photos==1.0
urlgrabber==3.10.2
urllib3==1.25.8
virtualenv==20.0.13
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
yum-metadata-parser==1.1.4
zipp==1.2.0
=======
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
protobuf==3.13.0
six==1.15.0
Werkzeug==1.0.1
>>>>>>> fffdde3ff50d457928c504a616f81e5631fd65e9
